+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Feature:
 
 - Show error msg on corrupted save file
 - Have login to steam save convert to no login to steam save (have app_id to no app_id)
+- Migrate player to specified Steam ID
 - MP loading progress
 - Delete Damange Object will delete no CharacterId object